phân quyền nhóm quyền

+++ controllers/admin/role.controller.ts
@@ -4,16 +4,22 @@ import { systemConfig } from "../../config/system";
 
 //[GET] / admin/roles
 export const index = async (req: Request, res: Response) => {
+    const permissions = res.locals.role.permissions;
 
-    const records = await Role.findAll({
-        where: { deleted: false },
-        raw: true
-    })
+    if (!permissions.includes("role_view")) {
+        res.status(403).send("Bạn không có quyền xem nhóm quyền");
+        return;
+    } else {
+        const records = await Role.findAll({
+            where: { deleted: false },
+            raw: true
+        })
 
-    res.render("admin/pages/roles/index", {
-        pageTitle: "Danh sách nhóm quyền",
-        records: records,
-    });
+        res.render("admin/pages/roles/index", {
+            pageTitle: "Danh sách nhóm quyền",
+            records: records,
+        });
+    }
 }
 
 //[GET] / admin/roles/create
@@ -25,31 +31,45 @@ export const create = async (req: Request, res: Response) => {
 
 //[ POST] / admin/roles/create
 export const createPost = async (req: Request, res: Response) => {
-    try {
-        await Role.create({
-            title: req.body.title,
-            description: req.body.description,
-        });
-        req.flash('success', 'Thêm nhóm quyền mới thành công');
-        res.redirect(`/${systemConfig.prefixAdmin}/roles`);
-    } catch (error) {
-        req.flash('error', 'Thêm nhóm quyền mới thất bại');
-        res.redirect(`/${systemConfig.prefixAdmin}/roles`);
+    const permissions = res.locals.role.permissions;
+
+    if (!permissions.includes("role_create")) {
+        res.status(403).send("Bạn không có quyền tạo nhóm quyền");
+        return;
+    } else {
+        try {
+            await Role.create({
+                title: req.body.title,
+                description: req.body.description,
+            });
+            req.flash('success', 'Thêm nhóm quyền mới thành công');
+            res.redirect(`/${systemConfig.prefixAdmin}/roles`);
+        } catch (error) {
+            req.flash('error', 'Thêm nhóm quyền mới thất bại');
+            res.redirect(`/${systemConfig.prefixAdmin}/roles`);
+        }
     }
 }
 
 //[GET] / admin/roles/detail/:id
 export const detail = async (req: Request, res: Response) => {
-    const id = req.params.id;
-    const record = await Role.findOne({
-        where: { id: id, deleted: false },
-        raw: true
-    });
-    res.render("admin/pages/roles/detail", {
-        // const record = await Role.findOne({ _id: req.params.id });
-        pageTitle: `Chi tiết ${record["title"]}`,
-        record: record,
-    });
+    const permissions = res.locals.role.permissions;
+
+    if (!permissions.includes("role_view")) {
+        res.status(403).send("Bạn không có quyền xem nhóm quyền");
+        return;
+    } else {
+        const id = req.params.id;
+        const record = await Role.findOne({
+            where: { id: id, deleted: false },
+            raw: true
+        });
+        record["permissions"] = JSON.parse(record["permissions"]);
+        res.render("admin/pages/roles/detail", {
+            pageTitle: `Chi tiết ${record["title"]}`,
+            record: record,
+        });
+    }
 }
 
 //[GET] / admin/roles/edit/:id
@@ -67,23 +87,49 @@ export const edit = async (req: Request, res: Response) => {
 
 //[PATCH] / admin/roles/edit/:id
 export const editPatch = async (req: Request, res: Response) => {
-    const id = req.params.id;
-    await Role.update({
-        title: req.body.title,
-        description: req.body.description,
-    }, {
-        where: { id: id }
-    })
-    req.flash('success', 'Cập nhật nhóm quyền thành công');
-    res.redirect(`/${systemConfig.prefixAdmin}/roles`);
+    const permissions = res.locals.role.permissions;
+
+    if (!permissions.includes("role_edit")) {
+        res.status(403).send("Bạn không có quyền chỉnh sửa nhóm quyền");
+        return;
+    } else {
+        try {
+            const id = req.params.id;
+            await Role.update({
+                title: req.body.title,
+                description: req.body.description,
+            }, {
+                where: { id: id }
+            })
+            req.flash('success', 'Cập nhật nhóm quyền thành công');
+            res.redirect(`/${systemConfig.prefixAdmin}/roles`);
+        } catch (error) {
+            req.flash('error', 'Lỗi');
+            res.redirect(`back`);
+        }
+
+    }
+
 }
 
 //[ DELETE] / admin/roles/delete/:id
 export const deleteRole = async (req: Request, res: Response) => {
-    const id = req.params.id;
-    await Role.destroy({ where: { id: id } });
-    req.flash('success', 'Xóa nhóm quyền thành công');
-    res.redirect(`/${systemConfig.prefixAdmin}/roles`);
+    const permissions = res.locals.role.permissions;
+
+    if (!permissions.includes("role_delete")) {
+        res.status(403).send("Bạn không có quyền xóa nhóm quyền");
+        return;
+    } else {
+        try {
+            const id = req.params.id;
+            await Role.destroy({ where: { id: id } });
+            req.flash('success', 'Xóa nhóm quyền thành công');
+            res.redirect(`/${systemConfig.prefixAdmin}/roles`);
+        } catch (error) {
+            req.flash('success', 'lỗi');
+            res.redirect(`back`);
+        }
+    }
 }
 
 //[GET] / admin/roles/permissions
@@ -100,21 +146,27 @@ export const permissions = async (req: Request, res: Response) => {
 
 //[Patch] / admin/roles/permissions
 export const permissionsPatch = async (req: Request, res: Response) => {
+    const permissions = res.locals.role.permissions;
 
-    try {
-        const permissions = JSON.parse(req.body.permissions);
+    if (!permissions.includes("role_permissions")) {
+        res.status(403).send("Bạn không có quyền cập nhật phân quyền");
+        return;
+    } else {
+        try {
+            const permissions = JSON.parse(req.body.permissions);
 
-        for (const item of permissions) {
-            await Role.update({
-                permissions: item.permissions
-            }, {
-                where: { id: item.id }
-            })
+            for (const item of permissions) {
+                await Role.update({
+                    permissions: item.permissions
+                }, {
+                    where: { id: item.id }
+                })
+            }
+
+            req.flash('success', `Cập nhật phân quyền thành công!`);
+        } catch (error) {
+            req.flash('error', `Cập nhật phân quyền thất bại!`);
         }
-
-        req.flash('success', `Cập nhật phân quyền thành công!`);
-    } catch (error) {
-        req.flash('error', `Cập nhật phân quyền thất bại!`);
+        res.redirect("back");
     }
-    res.redirect("back");
 }
